update snapshot db
